refactor: replace searchAPI with useSearchApi hook to prevent infinite loop rendering
